DINOV2 Base/requirement.txt
torch
torchvision
timm
numpy
Pillow
pandas
requests
pymongo
dnspython
fastapi
uvicorn
python-multipart
python-dotenv